refactor(tools): move git mock stuff back from utils due to "cfg!(test)"
we have to expose the "try_exec_git()" function.
