+++ crates/rari-tools/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod add_redirect;
 pub mod error;
 pub mod fix;
+pub mod git;
 pub mod history;
 pub mod inventory;
 pub mod r#move;
